// subject-sellcake-frontend-admin/src/app/_modules/model/user.ts
export class User {
  userId: string;
  firstName: string;
  lastName: string;
  fullName: string;
  email: string;
  userName: string;
  password: string;
  address: string;
  birthday: string;
  phone: string;
  imageFile: File;
  status: number;
  createDate: string;
  images: string;
  lastActivity: Date;
  lang: string;
  role: string;
}